requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
oauth2client==4.1.3
oauthlib==3.2.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pygame==2.6.1
pyparsing==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
python-dotenv
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
flask_login



